Removed exposed discord webhook
If you go to the discord webhook method that is inside the class called extra_shii, u will be able to place ur discord webhook that will be notified upon successfull bruteforce attempts.

Keep in mind this script is still under production and False positives are not uncommon. So please be sure to read server responses to validate True from False Positives.